refactor: simplify parsing of module graph (wip)
dependencies are now a list of pointers in the Module struct. the module list is now the single source of truth and we don't have to fiddle with multiple data structures anymore. this will also allow us to trivially modify module information and have it propagate through list and graph.

BREAKING CHANGE: replacements are not considered descendants anymore (#8)

Signed-off-by: nscuro <nscuro@protonmail.com>